Fixing Fabio's comments from PR w/ Fabio
